# Conflicts:
#	src/main/java/com/example/miniproject/service/UserService.java

+++ src/main/java/com/example/miniproject/service/UserService.java
@@ -36,7 +36,7 @@ public class UserService {
         System.out.println(user.getPassword());
 
         System.out.println(passwordEncoder.encode(password));
-        if (!passwordEncoder.matches(password,user.getPassword())) {
+        if (!passwordEncoder.matches(password, user.getPassword())) {
             throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
         }
 
@@ -59,7 +59,16 @@ public class UserService {
 
     @Transactional
     public void logout(HttpServletRequest request) {
-        redisService.deleteValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER));
+        if (request.getHeader(JwtUtil.REFRESHTOKEN_HEADER) == null) {
+            throw new IllegalStateException("잘못된 접근입니다");
+        }
+
+        if (redisService.getValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER)) != null) {
+            redisService.deleteValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER));
+        } else {
+            throw new IllegalStateException("잘못된 접근입니다");
+        }
+
     }
 
     //회원가입
